requirements.txt
selenium==4.25.0
requests==2.32.3
python-dotenv==1.0.1
